eighth lesson done

+++ app/Http/Controllers/PostController.php
@@ -10,6 +10,10 @@ use Illuminate\Support\Str;
 
 class PostController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth')->except('index', 'show');
+    }
     /**
      * Display a listing of the resource.
      *
@@ -53,7 +57,7 @@ class PostController extends Controller
         $post->short_title = Str::length($request->title) > 30 ? Str::substr($request->title, 0, 30) . '...' : $request->title;
         $post->descr = $request->descr;
 
-        $post->author_id = rand(1, 4);
+        $post->author_id = \Auth::user()->id;
 
         if ($request->file('img')) {
             $path = Storage::putFile('public', $request->file('img'));
@@ -74,6 +78,9 @@ class PostController extends Controller
     {
         $post = Post::join('users', 'author_id', '=', 'users.id')
             ->find($id);
+        if (!$post) {
+            return redirect()->route('post.index')->withErrors('Что Вы пытаетесь этим доказать?');
+        }
         return view('posts.show', compact('post'));
     }
 
@@ -86,6 +93,13 @@ class PostController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+        if (!$post) {
+            return redirect()->route('post.index')->withErrors('Что Вы пытаетесь этим доказать?');
+        }
+        if ($post->author_id != \Auth::user()->id) {
+            return redirect()->route('post.index')->withErrors('Вы не можете редактировать данный пост!');
+        }
+
         return view('posts.edit', compact('post'));
     }
 
@@ -99,6 +113,13 @@ class PostController extends Controller
     public function update(PostRequest $request, $id)
     {
         $post = Post::find($id);
+        if (!$post) {
+            return redirect()->route('post.index')->withErrors('Что Вы пытаетесь этим доказать?');
+        }
+        if ($post->author_id != \Auth::user()->id) {
+            return redirect()->route('post.index')->withErrors('Вы не можете редактировать данный пост!');
+        }
+
         $post->title = $request->title;
         $post->short_title = Str::length($request->title) > 30 ? Str::substr($request->title, 0, 30) . '...' : $request->title;
         $post->descr = $request->descr;
@@ -121,6 +142,13 @@ class PostController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+        if (!$post) {
+            return redirect()->route('post.index')->withErrors('Что Вы пытаетесь этим доказать?');
+        }
+        if ($post->author_id != \Auth::user()->id) {
+            return redirect()->route('post.index')->withErrors('Вы не можете удалить данный пост!');
+        }
+
         $post->delete();
         return redirect()->route('post.index')->with('success', 'Пост успешно удалён!');
     }
